fixed compiler errors

+++ badasher/Assets/Code/Objects/PlayerRelated/PlayerConstants.cs
@@ -21,8 +21,8 @@ public static class PlayerConstants {
 	public const float BOOST_POWER_INPUT_BUFFER = 0.2f; // in millisecondsS
 
 
-	public const float ENEMY_BP_DAMAGE = 20; // out of Boost Power Max
+	public const int ENEMY_BP_DAMAGE = 20; // out of Boost Power Max
 
 
-	public const float POWERUP_BOOST_AMOUNT = 40;
+	public const int POWERUP_BOOST_AMOUNT = 40;
 }
